steps_count made int
Also added more complex tests for it

+++ modules/task_1/vlasov_m_simpson_method/simpson_method.h
@@ -10,7 +10,7 @@ namespace SimpsonMethod {
 
 double integrate(const std::function<double(const std::vector<double>&)>& func,
                  const std::vector<double>& seg_begin,
-                 const std::vector<double>& seg_end, size_t steps_count);
+                 const std::vector<double>& seg_end, int steps_count);
 
 }  // namespace SimpsonMethod
